# [ArrowStringArray] PERF: small perf gain for object fallback