```
       before           after         ratio
     [a43c42c3]       [a5dd50ea]
     <master>         <small-perf>
-     26.8±0.05ms       26.1±0.1ms     0.97  strings.Methods.time_get('arrow_string')
-      21.2±0.1ms       20.3±0.1ms     0.96  strings.Methods.time_replace('arrow_string')
-      18.1±0.1ms       17.3±0.1ms     0.96  strings.Methods.time_lstrip('arrow_string')
-      18.5±0.1ms       17.7±0.2ms     0.95  strings.Methods.time_strip('arrow_string')
-      18.3±0.1ms      17.4±0.09ms     0.95  strings.Methods.time_rstrip('arrow_string')
-      28.0±0.3ms      26.6±0.07ms     0.95  strings.Methods.time_center('arrow_string')
-      21.2±0.2ms       20.1±0.1ms     0.95  strings.Methods.time_normalize('arrow_string')
-      28.1±0.2ms       26.5±0.2ms     0.94  strings.Methods.time_pad('arrow_string')
-      13.6±0.1ms       11.5±0.1ms     0.85  strings.Methods.time_isdigit('str')
-      13.6±0.2ms      11.3±0.04ms     0.83  strings.Methods.time_isnumeric('str')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```